util: refactor render_raw_template_string to tim_common module
